ci: Upload iso images from MRs as well
This allows running test-s3-image (the openQA end-to-end tests) on all
branches, including open merge requests. Pre-merge openQA testing!

Replaces GNOME/gnome-build-meta!2356

Closes #601


Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3949>

+++ .gitlab-ci/scripts/publish-s3-image.sh
@@ -1,13 +1,19 @@
 #! /bin/bash
 
-set -euox pipefail
+set -euo pipefail
 
-if [[ -n "${CI_COMMIT_TAG:-}" ]]; then
-    aws s3 cp --acl public-read live-image/disk.iso \
-        "s3://gnome-build-meta/tag/${CI_COMMIT_TAG}/live_${CI_COMMIT_TAG}-${ARCH}.iso"
+
+if [ "${CI_COMMIT_REF_PROTECTED:-}" = true ]; then
+    if [[ -n "${CI_COMMIT_TAG:-}" ]]; then
+        aws s3 cp --acl public-read live-image/disk.iso \
+            "s3://gnome-build-meta/tag/${CI_COMMIT_TAG}/live_${CI_COMMIT_TAG}-${ARCH}.iso"
+    else
+        aws s3 cp --acl public-read live-image/disk.iso \
+            "s3://gnome-build-meta/nightly/${CI_PIPELINE_ID}/live_${CI_PIPELINE_ID}-${ARCH}.iso"
+    fi
 else
     aws s3 cp --acl public-read live-image/disk.iso \
-        "s3://gnome-build-meta/nightly/${CI_PIPELINE_ID}/live_${CI_PIPELINE_ID}-${ARCH}.iso"
+        "s3://gnome-build-meta/merge_request/mr${CI_MERGE_REQUEST_IID}/live_mr_${CI_PIPELINE_ID}-${ARCH}.iso"
 fi
 
 aws s3 ls --recursive --human-readable s3://gnome-build-meta/
